pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "zzag-narratisedb"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=6.2.1,<7.0.0)",
    "supabase (>=2.18.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
autopep8 = "^2.3.2"
pytest = "^8.4.1"

[tool.flake8]
max-line-length = 160

